refactor: seperate apiCall
